It has been decided to stop implementing recursive analyzer the way they are defined in IDMEF since it prevent to do any kind of 'dynamic' filtering since we don't know the element depth. Rather we now implement analyzer as a list, and provide compatibility when exporting to XML. This is the solution we came up with when talking with one of the IDWG chair.
Also, modify list related function so that they take a 'pos' argument. You can now say
whether the object should be positionned at the beginning, the end, or any relative
position. 

Modify the _new_child() function used by idmef-path so that it error out if an user ask
to create an index for which lower index have not been created. This avoid creating as 
much empty index as the user specified.



git-svn-id: file:///home/yoann/SVN/trunk/libprelude@5392 09c5ec92-17d4-0310-903a-819935f44dba