New substr operator for substr : "<>". Provide the ability to do insensitive match by appending a wilcard to most (<>, =, ~) operator. Provide the ability to do negative match on <> and ~ operator. Implement lesser and greater operator for broken down time operation.
Make the operator enum a bitsflags which make the code a little more easier to handle.
Ability for idmef-criterion-value to print regex.

Fix some bug in broken down time month/wday handling. Ability to provide a gmt offset
in the criteria so that the match is not made on the local time.

Watch out, libpreludedb breakage.


git-svn-id: file:///home/yoann/SVN/trunk/libprelude@6968 09c5ec92-17d4-0310-903a-819935f44dba