When compiling against recent GnuTLS version, don't link prelude-adduser to libgnutls-extra since, SRP symbols were moved to libgnutls.
If no SRP symbol can be found in either gnutls or gnutls-extra, abort with
an error explaining the user what's going on:

"Your operating system distribution provides the GnuTLS library.
Unfortunately, the provided binary lacks support for the SRP protocol.

If you are seeing this message, it probably mean you're running on Redhat Fedora, 
which distributes a stripped down version of the GnuTLS library. 
Unfortunately, removing part of a library's API break certain applications relying on the library.

Solutions:
	- Install GnuTLS from http://www.gnutls.org
	- File a Redhat ticket on https://bugzilla.redhat.com/"


git-svn-id: file:///home/yoann/SVN/trunk/libprelude@7345 09c5ec92-17d4-0310-903a-819935f44dba