Fix float comparison issue reported by Steve Grubb <sgrubb@redhat.com>.
The fix introduce a new float_compare() function which subtract the
integer representations of two floats to check how close they are.
If the difference is zero, they are identical. If the difference is
one, they are adjacent floats. In general, if the difference is n
then there are n-1 floats between them.

git-svn-id: file:///home/yoann/SVN/libprelude/trunk@11127 09c5ec92-17d4-0310-903a-819935f44dba