Improve retrieval of list of IDMEF object
When retrieving a list of IDMEF object, return a Python list containing
one IDMEFValue for each object.

This behavior is more consistant than retrieving a single IDMEFValue
carrying the whole list, and might permit the user to play with the
object directly if we extand the API in the future.

Compatibility with the old implementation for setting whole list is
achived by converting back the list of IDMEFValue to an IDMEFValue
of the list type on Set().