Prevent IDMEF copy from failing if a required children is unavailable
The previous behavior when copying an IDMEF object was to fail if
one of the children to be copied was required but unavailable.

Since we don't want to make assumption in the API (the caller could
temporarily destroy a required children, to recreate it later), carry
on with the copy if one of the required children is not available.