Fix high level bindings ambiguity on IDMEFPath.Set() and IDMEF.Set() methods
The dispatch function generated by SWIG is not capable of distinguishing
between a string consisting of a single character, and a int8/uint8 integer.

As a result, when setting an IDMEFPath to a value containing a single
character (example: idmef.Set("alert.additionnal_data(0).data", "c")), the
generated path value was configured as carrying an integer, rather than a
string.

As a result, we ignore the C++ overloaded function for int8/uint8 type. If
bigger integer are passed to IDMEF path that only support 8 bits integers,
libprelude will perform value checking and cast the value if required.