Comparison operator support for IDMEFValue()
Provide equal, not equal, lesser than, greater than operator so that it
is possible to directly compare two value together, without calling the
IDMEFValue.Match() method.

Additionally, make it possible to call the equal operator with a Vector
of IDMEFValue() as argument, which make it possible to support Python
"in" operator (item in valuelist).

Change-Id: Ic5dcb1372183f580b431cd21a25ce38d695d63c1