Improved error handling, fix 8 bits unsigned integer conversion to string
Converting a 8 bits unsigned integer to a string could result in an
invalid value, because of an invalid unsigned to signed cast.

Additionally, when parsing integer, some error were not correctly
detected (for example left over character in the input string).

This change fixes the problem, and should also improve performance by
making use of strto* functions in place of sscanf().

This fixes an issue in libpreludedb/Prewikka, where invalid IDMEFCriteria
were not correctly reported, leading to a later libpreludedb error.

Fix #546.

Change-Id: I07b800b48687f511fb629c4fe8f00584b46d170b