Fix asynchronous stack / high level bindings callback deadlock
We need to unlock the interpreter lock before calling certain methods
because they might acquire internal libprelude mutex that may also be
acquired undirectly through the libprelude asynchronous stack.

[Thread 2]: Libprelude async stack
-> Lock internal mutexX
   -> prelude_log()
      -> SWIG/C log callback
         -> Wait on Interpreter lock [DEADLOCK]
            -> Python logging callback (never called)

[Thread 1] ConnectionPool::Recv()
-> Acquire Interpreter lock
   *** At this time, thread 2 lock internal mutexX
   -> Wait on internal mutexX [DEADLOCK]

In this situation, [Thread 1] hold the Interpreter lock and is
waiting on mutexX, which itself cannot be released by [Thread 2]
until [Thread 1] unlock the Interpreter lock.

One rule to prevent deadlock is to always acquire mutex in the same
order. We thus need to make sure the interpreter lock is released
before calling C++ method that are susceptible to lock a mutex that
is also triggered from the asynchronous interface.

Note that we are not releasing the Interpreter lock in all C++ call,
because it come at a performance cost, so we only try to do it when
needed.

Fix #331.

Change-Id: Iafd6d24145cf0281871b170701effcfe49c23af1