Implement prelude_hash_add()
The prelude_hash_set() function destroy any previously element with the
same key already in the hash.

We now additionally implement prelude_hash_add(), which add the new element
even if another element with the same key exist.

This is a micro-optimisation for case where the user never add element
that are already in the hash.

Change-Id: Ifc092a4e9f9698bf9ba936870e7665f98020e9da