Allow creation of IDMEFPath relative to a root
This introduce the idmef_path_new_from_root_fast() function, which allow
creating IDMEFPath relative to a specific root (rather than IDMEFMessage).

As an example :
 idmef_path_new_from_root_fast(&path, IDMEF_CLASS_ID_ANALYZER, "name");

Will create an IDMEFPath that reference the 'nam'e attribute of an
idmef_analyzer_t object.

The idmef_path_set() and idmef_path_get() function are modified so that
they can be called with any root IDMEF object as parameter (rather than
idmef_message_t).

The idmef_criteri(a|on) API is also modified so that it can be passed
any root object (since it use the idmef_path_t API internally).

Change-Id: I1bcacca83f8b4d9141c0455ee6c6d19ba0c218dd