Modify IDMEF and IDMEFPath C++ API to understand relative path
The IDMEF() object used to represent the root of an IDMEF message
(idmef_message_t), and is modified so that it now abstract any
IDMEF low level object.

When used to access an object rather than a value, the IDMEFPath.Get()
and IDMEF.Get() methods now directly return an IDMEF() object in place of
an IDMEFValue() representing the object.

Since relative path are now supported, it is now possible write :
- idmef.Get("alert").Get("analyzer(0)").Get("name")
- idmef.Get("alert").Get("source(0)").Get("node.address.address")

For the Python language, __getitem__ is implemented for IDMEF object,
so that it is possible to write the above as:

- idmef["alert"]["analyzer(0)"]["name"]
- idmef["alert"]["source(0)"]["node.adress.address"]

Change-Id: I8768cd78bc444f4fb580bf9432d70782fc6518da