prelude-linked-object members ordering fixes, so that the members are synchronized with listed idmef_object_t.
Implement idmef_class_is_listed(), to check whether a given IDMEF class
is a list element. IDMEFTreeData is modified as to include the information
about whether a given class is listed.

Implement list operation to the idmef_object_t abstraction. The functions
will print an assertion warning it case they are called on an idmef_object_t
element that isn't listed.

Make the prelude_string structure private, as it is supposed to be.
Additionally, the prelude_string_t object was using a regular list declaration,
in place of PRELUDE_LINKED_OBJECT. Nobody could notice until now,
because the list member was first in both place, but it would now
result in a bug since we changed the PRELUDE_LINKED_OBJECT members
order.

Modify idmef-tree-wrap generation so that it use prelude_linked_object_t
abstraction, since prelude-string structure is now private.

Change-Id: I78f64eb3ee202c7702bc31eb913565853a7d2f99