[#rel1574] broken-down time comparison worked only once
The first call to idmef_criteria_match() on a broken-down time criteria was
working, but subsequent call where not.

The reason is that we used timegm() on the internal TM structure used to hold
the comparison value. We need to perform a copy of this structure since
timegm() will modify it.

Change-Id: I5a4c5ffded185458be8694f4ceef4a864331775f